requirements.txt
docker-py==1.5.0
dnspython3
kazoo
